moving Logins to display/fewlines folder
